feat(game): weekly NPC turn-spender cron + /api/game/npc-weekly
Wires the NPC weekly turn-spender into a Sunday 05:30 UTC cron — 30 minutes
after the existing rollover, so any rollover state changes have settled
before NPCs run. No Firebase Admin credentials are exposed to GitHub Actions:
the workflow only knows the deployed endpoint URL and the shared rollover
secret; all credentials stay on Vercel.

Changes:
  - lib/game/npc-weekly.ts: extracts the script's main loop into a reusable
    runNpcWeeklyServer({weekStartIso, dryRun, limit}) so the route and the
    CLI share the same implementation. Returns NpcWeeklySummary for the API
    response. Idempotent per (player × weekStartIso).
  - app/api/game/npc-weekly/route.ts: cron-only POST, x-rollover-secret
    auth, maxDuration=300 to cover 50 NPCs × ~25 actions of Firestore txns.
  - .github/workflows/game-npc-weekly.yml: 30 5 * * 0 cron + manual dispatch
    with optional weekStartIso override and dryRun flag. Calls the endpoint
    with the existing GAME_ROLLOVER_SECRET; needs a new GAME_NPC_WEEKLY_URL
    secret pointing at the deployed route.
  - scripts/run-npc-weekly.ts: slimmed to a CLI wrapper around the server
    function for backfill / ad-hoc runs.
